pyproject.toml
[project]
name = "initial-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain==1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
